+++ packages/expo-modules-core/CHANGELOG.md
@@ -18,7 +18,7 @@
 - Fix failing instrumentation tests in JavaScriptViewModule. ([#22518](https://github.com/expo/expo/pull/22518) by [@aleqsio](https://github.com/aleqsio))
 - Fixed Android build warnings for Gradle version 8. ([#22537](https://github.com/expo/expo/pull/22537), [#22609](https://github.com/expo/expo/pull/22609) by [@kudo](https://github.com/kudo))
 - [Android] Fix the `Color` converter doesn't work on devices with SDK version below 26. ([#22191](https://github.com/expo/expo/pull/22191) by [@lukmccall](https://github.com/lukmccall))
-- Refactored network logger code and add unit tests on iOS. ([#22669](https://github.com/expo/expo/pull/22669) by [@kudo](https://github.com/kudo))
+- Refactored network inspector code and add unit tests on iOS. ([#22669](https://github.com/expo/expo/pull/22669) by [@kudo](https://github.com/kudo))
 
 ### 💡 Others